requirements.txt
# Core Libraries
numpy>=1.21.0,<2.0.0
scipy>=1.8.0,<1.14.0
scikit-learn>=1.0.0
torch
tqdm

# Database Libraries
faiss-cpu
# faiss-gpu - Use conda for GPU
pymilvus
pinecone
qdrant-client
weaviate-client

# LLM Integration
transformers
ollama
groq
openai

# Retrieval-Augmented Generation (RAG)
matplotlib
tabula-py
camelot-py
pymupdf
sentence-transformers
tensorflow_hub
gensim
layoutparser
pdf2image
pytesseract
pdfplumber
#spacy
fastapi[standard]
jpype1
llama-index-core==0.12.2
llama-parse
llama-index-readers-file
python-dotenv
opencv-python
datasets
pptx
moviepy
speech_recognition
ebooklib
bs4
docx
# NLP Models
# Uncomment the following line to download Spacy's English model
# python -m spacy download en_core_web_sm
